+++ src/plugin.ts
@@ -1,0 +1,6 @@
+import {type Plugin} from '@commitlint/types';
+import {rules} from './rules.js';
+
+const plugin: Plugin = {rules};
+
+export {plugin};